Hopefully fixed no power off bug
I moved the power latch later in the setup sequence - my theory is that
the chip was browning out when the power pin brought the power down,
jumping to the reset vector (without fully resetting), and running main
again. Since the (nearly) first instruction in main was to latch the
power on, the power stayed on. Moving the latch instruction to right
before the first buffer fill should still be quick enough that a human
button press will be latched but slow enough that a brownout partial
reset cannot re-latch the power. Only testing will tell.......