testing random function
It's working - the reason it seemed like it was locking up was that one
of my filenames was too long; I fixed that and it hits every time now.
It's still a little too 'consistent' for my liking, but humans aren't
too good at true randomness so <shrug>. I'll search around for other
ways to generate 'randomness' on an AVR, but if I need to stick with
this, I'll just artificially weight the sounds I like better by putting
them in the table multiple times.